Combine exec and server modes into adhoc mode.
A new command adhoc is created that combines exec and server modes
into a single process. The approach is quite hacky, but it works:
- the server and the agent are run in a single process.
- once the agent finishes, the server is still running and allows to
visualize the data in the normal UI.

On the other hand, a lot of things are still missing:
- Only exec mode targets are supported.
- The agent uses the Upstream mode instead of Direct one.
- The configuration is also a combination of the exec and server
configurations, which is unnecessarily complex and presents some
problems. Depending the profiling data wants to be presented, many of
these configuration options will be unnecesary.
- The file-system is used to store the data, instead of using memory.