Polish adhoc mode.
Several new features are implemented:
- The agent uses direct upstream instead of remote upstream.
- The server doesn't upstream its own data.
- In memory badger is used for the profiling data.
- The unneede CLI options are removed.

From the user point of view, these features give a reasonable approach
to the first step of adhoc mode, from the execution and ingestion
point of view, but persisting the generated data is still pending.

The solution has some shortcomings codewise. In order to support the
particularities of the adhoc mode, some dependency construction has
been moved outside of the components that use them. This is good, but
there's still quite a lot of coupling between different parts, and as
a result the way this extractions work now has room for improvement.
This include:
- The logger: it's constructed and then passed as a dependency to the
storage, server, and exec components. This is an attempt to unify how
logging is handled, but it's still not completely consistent.
- The storage: it's also constructed and then passed as a dependency
to the server and exec components. The rest of the exec modes don't
need a storage, but since there's no constructor for the exec, the way
to handle this is currently hacky. It may be better to actually split
the whole exec component by mode, but there would be also some
duplication that is not easy to isolate.